Start adding support for Mitsubishi MELPS-740
The MELPS-740 family is a 6502 variant with some extra instructions and
addressing modes.

The "T" flag (use memory in X register instead of ACC) is not yet
implemented (only the flag has been added).

Divide and multiply similarly are not implemented.

The new indirect jump and JSR instructions need testing.